Refactoring of NpmDependenciesManipulator name and creation of dedicated test class
